Moving field functions in their own location
Closes #29
