+++ framework/field_functions/interpolation/ffinter_slice.cc
@@ -1,11 +1,11 @@
-#include "framework/mesh/field_function_interpolation/ffinter_slice.h"
+#include "framework/field_functions/interpolation/ffinter_slice.h"
 #include "framework/mesh/cell/cell.h"
 #include "framework/mesh/raytrace/raytracer.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/math/spatial_discretization/spatial_discretization.h"
 #include "framework/math/vector_ghost_communicator/vector_ghost_communicator.h"
 #include "framework/math/spatial_discretization/finite_element/quadrature_point_data.h"
-#include "framework/physics/field_function/field_function_grid_based.h"
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
 #include <fstream>
