+++ framework/field_functions/interpolation/ffinter_volume.cc
@@ -1,5 +1,5 @@
-#include "framework/mesh/field_function_interpolation/ffinter_volume.h"
-#include "framework/physics/field_function/field_function_grid_based.h"
+#include "framework/field_functions/interpolation/ffinter_volume.h"
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/math/spatial_discretization/spatial_discretization.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
 #include "framework/math/vector_ghost_communicator/vector_ghost_communicator.h"
