+++ framework/field_functions/operations/multi_field.cc
@@ -1,7 +1,5 @@
-#include "framework/physics/field_operations/multi_field.h"
-
+#include "framework/field_functions/operations/multi_field.h"
 #include "framework/object_factory.h"
-
 #include "framework/math/spatial_discretization/spatial_discretization.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
