+++ framework/field_functions/operations/partitioner_predicate.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "framework/physics/field_operations/field_operation.h"
+#include "framework/field_functions/operations/field_operation.h"
 
 namespace opensn
 {
