+++ lua/framework/physics/field_function/physics_export_field_func.cc
@@ -1,7 +1,5 @@
 #include "framework/lua.h"
-
-#include "framework/physics/field_function/field_function_grid_based.h"
-
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
 #include "field_functions_lua.h"
