+++ lua/framework/physics/field_function/physics_get_field_func_list.cc
@@ -1,10 +1,7 @@
 #include "framework/lua.h"
-
 #include "framework/runtime.h"
-
 #include "framework/physics/solver_base/solver.h"
-#include "framework/physics/field_function/field_function_grid_based.h"
-
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/logging/log.h"
 #include "field_functions_lua.h"
 #include "framework/console/console.h"
