+++ lua/modules/linear_bolzmann_solvers/lbs_solver/lbs_get_field_funcs.cc
@@ -1,7 +1,5 @@
 #include "modules/linear_boltzmann_solvers/a_lbs_solver/lbs_solver.h"
-
-#include "framework/physics/field_function/field_function_grid_based.h"
-
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
 #include "framework/lua.h"
