+++ modules/linear_boltzmann_solvers/a_lbs_solver/lbs_solver.cc
@@ -19,7 +19,7 @@
 #include "framework/math/spatial_discretization/finite_element/piecewise_linear/piecewise_linear_discontinuous.h"
 #include "framework/physics/physics_material/physics_material.h"
 #include "framework/physics/physics_material/multi_group_xs/adjoint_mgxs.h"
-#include "framework/physics/field_function/field_function_grid_based.h"
+#include "framework/field_functions/field_function_grid_based.h"
 #include <algorithm>
 #include <iomanip>
 #include <sys/stat.h>
