+++ test/framework/tutorials/fv_test1.cc
@@ -1,15 +1,10 @@
 #include "framework/mesh/mesh_handler/mesh_handler.h"
-
 #include "framework/math/spatial_discretization/finite_volume/finite_volume.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
-
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
-
-#include "framework/physics/field_function/field_function_grid_based.h"
-
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
-
 #include "lua/framework/console/console.h"
 
 using namespace opensn;
