+++ test/modules/linear_boltzmann_solvers/dsa/acceleration_diffusion_dfem.cc
@@ -1,17 +1,12 @@
 #include "framework/mesh/mesh_handler/mesh_handler.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
-
 #include "framework/math/spatial_discretization/finite_element/piecewise_linear/piecewise_linear_discontinuous.h"
-
 #include "modules/linear_boltzmann_solvers/a_lbs_solver/acceleration/acceleration.h"
 #include "modules/linear_boltzmann_solvers/a_lbs_solver/acceleration/diffusion_mip_solver.h"
 #include "modules/linear_boltzmann_solvers/a_lbs_solver/lbs_structs.h"
-
-#include "framework/physics/field_function/field_function_grid_based.h"
-
+#include "framework/field_functions/field_function_grid_based.h"
 #include "framework/runtime.h"
 #include "framework/logging/log.h"
-
 #include "lua/framework/console/console.h"
 #include "lua/framework/math/functions/lua_scalar_spatial_function.h"
